refactor pull request view and view builder
